Corrected issues when new models + elasticity: - data converter now sets backs -999999 to nan - avoided divisions by zeros in aircraft efficiency and energy use choices - small dtype issue (object type introduced by covid effect function) causing plot to fail. Do not understand why it was working in both branches...?

+++ aeromaps/models/air_transport/air_traffic/price_elasticity.py
@@ -83,7 +83,9 @@ class RPKWithElasticity(AeroMAPSModel):
             covid_end_year_reference_rpk_ratio / 100,
         ]
         covid_function = interp1d(reference_years, reference_values_covid, kind="linear")
-        covid_factors = pd.Series([covid_function(k) for k in covid_years], index=covid_years)
+        covid_factors = pd.Series(
+            [float(covid_function(k)) for k in covid_years], index=covid_years, dtype=float
+        )
 
         # CAGR function
         ## Short range
